Moved linked references to image from imageDetailAdminPage to ImageForm
